move execution lifecycle hooks setup out of workflow-execute-additional-data
