+++ packages/cli/src/execution-lifecycle-hooks/__tests__/execution-lifecycle-hooks.test.ts
@@ -28,7 +28,7 @@ import {
 	getWorkflowHooksMain,
 	getWorkflowHooksWorkerExecuter,
 	getWorkflowHooksWorkerMain,
-} from '../workflow-execute-additional-data';
+} from '../execution-lifecycle-hooks';
 
 describe('Execution Lifecycle Hooks', () => {
 	mockInstance(Logger);
