+++ packages/cli/src/executions/execution-recovery.service.ts
@@ -9,9 +9,9 @@ import { ExecutionRepository } from '@/databases/repositories/execution.reposito
 import { NodeCrashedError } from '@/errors/node-crashed.error';
 import { WorkflowCrashedError } from '@/errors/workflow-crashed.error';
 import { EventService } from '@/events/event.service';
+import { getWorkflowHooksMain } from '@/execution-lifecycle-hooks/execution-lifecycle-hooks';
 import type { IExecutionResponse } from '@/interfaces';
 import { Push } from '@/push';
-import { getWorkflowHooksMain } from '@/workflow-execute-additional-data'; // @TODO: Dependency cycle
 
 import type { EventMessageTypes } from '../eventbus/event-message-classes';
